platform: add a GCC specific stack smashing protection support
The GCC supports a "-fstack-protector" flag, which puts a canary
to stack and its validation before function return. This makes
it possible to detect stack overflows, such as in example below:

--8<--8<--
int make_it_crash(void)
{
    char too_small[8];
    strcpy(too_small, "too-big!");
    return 0;
}
--8<--8<--

The code below will create a system error and a log line such as:
"** stack smashing detected at 0x53b39 ***"

Note: the code needs to be compiled with a profile which has it,
such as the attached debug_size_stack_protector, which also
enables fist level optimizations, which are really needed as the
binary size grows with these checks.